+++ data/src/main/java/com/linku/data/di/preference/NotificationPreferenceModule.kt
@@ -2,7 +2,7 @@ package com.linku.data.di.preference
 
 import android.content.Context
 import com.linku.data.implementation.preference.NotificationPreferenceImpl
-import com.linku.data.preference.NotificationPreference
+import com.linku.core.repository.NotificationPreference
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
